apps/opportunity_finder/embedding_service/requirements.txt
openai==1.3.5
qdrant-client==1.7.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
numpy==1.24.3
sentence-transformers==2.3.1
torch==2.2.0
safetensors==0.4.1
httpx==0.25.2
asyncio-throttle==1.0.2